+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - FIX(sqlite) use dollar params instead of named params
 - CHANGE run performance tests without the `dev-mode` plugin
 - IMPROVE performance of document writes by not using try-catch in a hot path.
+- FIX `RxDatabase.remove()` must properly remove the collection storage together with the replication states.
 
 <!-- ADD new changes here! -->